Consolidate in-progress working tree
Snapshot of the accumulated multi-feature work on top of the GPT-5.6 /
per-model-context-window commits, so it can be merged to main and the dev
app recompiled with everything in place. Major areas:

- Home page redesign: usage heatmap (GitHub-style contribution graph),
  token-accounting audit (atomic i64 accumulation, token_events table +
  schema v10), GET /sessions/activity route, and removal of the flat
  session/token stat tiles (heatmap carries the usage story; recent chats
  expanded to 6). Heatmap tooltip portalled to <body> so it can't escape
  the window; streak heading sized down for hierarchy.
- Knowledge section redesign onto the design system: round-by-role radii,
  flat surfaces (no ad-hoc tints/shadows), new shared Badge primitive,
  input-shaped selects, hairline rows.
- Design system: design.md, docs/design-system.html, homepage- and
  knowledge-redesign proposals.
- Bedrock / Versa-Bedrock error classification + operation timeout.
- Agent Drafter apps (appcontrol ui_* tools, runnable export, examples).
- Auto Visualiser dashboard/report refinements.
- Assorted desktop UI alignment (icons, dashboard, workflows, skills,
  onboarding), FlyingBird removal.

+++ crates/biorouter-server/src/routes/schedule.rs
@@ -66,9 +66,9 @@ pub struct SessionDisplayInfo {
     total_tokens: Option<i32>,
     input_tokens: Option<i32>,
     output_tokens: Option<i32>,
-    accumulated_total_tokens: Option<i32>,
-    accumulated_input_tokens: Option<i32>,
-    accumulated_output_tokens: Option<i32>,
+    accumulated_total_tokens: Option<i64>,
+    accumulated_input_tokens: Option<i64>,
+    accumulated_output_tokens: Option<i64>,
 }
 
 #[utoipa::path(
